Move the create method of ExtendedRCG to PartialCG
